+++ src/Utility/IMAPFolderCleanupHelper.php
@@ -6,7 +6,6 @@ use DateInterval;
 use DateTimeImmutable;
 use Ddeboer\Imap\Search\Date\Before as SearchDateBefore;
 use Ddeboer\Imap\Server;
-use Foodsharing\Modules\Console\ConsoleControl;
 
 use function Sentry\captureException;
 
@@ -15,7 +14,7 @@ class IMAPFolderCleanupHelper
     public function cleanupFolder(?string $imapHost, ?string $imapUser, ?string $imapPass, ?string $folder, ?int $deleteDelayDays): int
     {
         if ($imapHost === null || $imapUser === null || $imapPass === null || $folder === null || $deleteDelayDays === null) {
-            ConsoleControl::error('Invalid parameters: All parameters must be provided.');
+            ConsoleHelper::error('Invalid parameters: All parameters must be provided.');
 
             return -1;
         }
@@ -27,7 +26,7 @@ class IMAPFolderCleanupHelper
 
             $mailbox = $connection->getMailbox($folder);
         } catch (\Throwable $e) {
-            ConsoleControl::error('Something went wrong connecting to folder "' . $folder . '", ' . $e->getMessage() . '\n');
+            ConsoleHelper::error('Something went wrong connecting to folder "' . $folder . '", ' . $e->getMessage() . '\n');
 
             return -1;
         }
@@ -47,7 +46,7 @@ class IMAPFolderCleanupHelper
                 ++$deleted;
             }
         } catch (\Throwable $e) {
-            ConsoleControl::error('Something went wrong removing old mails from "' . $folder . '", ' . $e->getMessage() . '\n');
+            ConsoleHelper::error('Something went wrong removing old mails from "' . $folder . '", ' . $e->getMessage() . '\n');
             captureException($e);
         } finally {
             $connection->expunge();